VeriDoc-RL/requirements/autodl.rl.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-cors==0.8.1
aiosignal==1.4.0
airportsdata==20260315
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asttokens==3.0.1
attrs==26.1.0
blobfile==3.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
codetiming==1.4.0
colorful==0.5.8
compressed-tensors==0.14.0.1
cryptography==46.0.5
cuda-bindings==13.2.0
cuda-pathfinder==1.4.3
cuda-python==13.2.0
datasets==4.8.3
decorator==5.2.1
dill==0.4.1
diskcache==5.6.3
distlib==0.4.0
einops==0.8.2
executing==2.2.1
fastapi==0.135.1
filelock==3.25.2
flashinfer-python==0.2.5+cu124torch2.6
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-auth==2.49.1
googleapis-common-protos==1.73.0
grpcio==1.78.0
h11==0.16.0
hf_transfer==0.1.9
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
interegular==0.3.3
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.3.1
librt==0.8.1
llguidance==0.7.30
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
modelscope==1.35.1
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
multiprocess==0.70.19
mypy==1.19.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
ninja==1.13.0
numpy==2.3.5
nvidia-ml-py==13.595.45
omegaconf==2.3.0
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.40.0
opentelemetry-exporter-prometheus==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
outlines==0.1.11
outlines_core==0.1.26
packaging==26.0
pandas==3.0.1
parso==0.8.6
partial-json-parser==0.2.1.1.post7
pathspec==1.0.4
peft==0.18.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.9.4
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.4.1
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybind11==3.0.2
pycountry==26.2.16
pycparser==3.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pylatexenc==2.10
pynvml==13.0.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-discovery==1.2.0
python-multipart==0.0.22
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.54.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rpds-py==0.30.0
ruff==0.15.7
safetensors==0.7.0
sentencepiece==0.2.1
sentry-sdk==2.55.0
setproctitle==1.3.7
sgl-kernel==0.1.4
sglang==0.4.6.post5
six==1.17.0
smart_open==7.5.1
smmap==5.0.3
soundfile==0.13.1
stack-data==0.6.3
starlette==0.52.1
sympy==1.13.1
tensordict==0.6.2
tiktoken==0.12.0
tokenizers==0.22.2
torchao==0.9.0
torchdata==0.11.0
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
triton==3.2.0
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
verl==0.4.1
virtualenv==21.2.0
wandb==0.25.1
wcwidth==0.6.0
wheel==0.46.3
wrapt==2.1.2
xgrammar==0.1.19
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
